+++ libs/ledger-live-common/src/modularDrawer/modules/createNetworkConfiguration.ts
@@ -1,5 +1,5 @@
-import type { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { getCryptoCurrencyById } from "@domain/entity-currency-crypto";
+import type { CryptoOrTokenCurrency } from "@domain/entity-currency";
 import type { ReactElement } from "react";
 import { useLeftAccountsModule } from "../hooks/useLeftAccounts";
 import { useLeftAccountsApyModule } from "../hooks/useLeftAccountsApy";
@@ -77,9 +77,9 @@ export function useNetworkConfiguration(
     const asset =
       network.type === "TokenCurrency" ? getCryptoCurrencyById(network.parentCurrencyId) : network;
     return {
-      ...asset,
       ...rightResults[rightElement]?.[i],
       ...leftResults[leftElement]?.[i],
+      ...asset,
     };
   });