+++ libs/ledger-live-common/src/modularDrawer/utils/index.ts
@@ -1,4 +1,4 @@
-import { CryptoOrTokenCurrency } from "@ledgerhq/types-cryptoassets";
+import { CryptoOrTokenCurrency } from "@domain/entity-currency";
 export { groupCurrenciesByAsset } from "./groupCurrenciesByAsset";
 export { sortAccountsByFiatValue } from "./sortAccountsByFiatValue";